ProMan Deploy/requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
Werkzeug==2.0.1
requests==2.9.1
gunicorn==19.7.1
python-dotenv==0.18.0
bcrypt==3.2.0